refactor (again): keep all lithostitched nodes in integration pkg
