+++ c2me-opts-dfc/src/main/java/com/ishland/c2me/opts/dfc/common/ast/opto/passes/FoldConstants.java
@@ -32,10 +32,10 @@ import com.ishland.c2me.opts.dfc.common.ast.binary.MaxShortNode;
 import com.ishland.c2me.opts.dfc.common.ast.binary.MinNode;
 import com.ishland.c2me.opts.dfc.common.ast.binary.MinShortNode;
 import com.ishland.c2me.opts.dfc.common.ast.binary.MulNode;
+import com.ishland.c2me.opts.dfc.common.ast.integration.lithostitched.unary.*;
 import com.ishland.c2me.opts.dfc.common.ast.misc.CacheLikeNode;
 import com.ishland.c2me.opts.dfc.common.ast.misc.ConstantNode;
 import com.ishland.c2me.opts.dfc.common.ast.unary.*;
-import com.ishland.c2me.opts.dfc.common.ast.unary.integration.lithostitched.*;
 import com.ishland.c2me.opts.dfc.common.util.ZeroUtils;
 import net.minecraft.util.math.MathHelper;
