+++ c2me-opts-dfc/src/main/java/com/ishland/c2me/opts/dfc/common/gen/jvm/emitters/UnaryNodeBytecodeEmitters.java
@@ -24,8 +24,8 @@
 
 package com.ishland.c2me.opts.dfc.common.gen.jvm.emitters;
 
+import com.ishland.c2me.opts.dfc.common.ast.integration.lithostitched.unary.*;
 import com.ishland.c2me.opts.dfc.common.ast.unary.*;
-import com.ishland.c2me.opts.dfc.common.ast.unary.integration.lithostitched.*;
 import com.ishland.c2me.opts.dfc.common.gen.CodeGenRegistry;
 import com.ishland.c2me.opts.dfc.common.gen.jvm.BytecodeEmitter;
 import com.ishland.c2me.opts.dfc.common.gen.jvm.BytecodeGen;
